servers/pw-mcp-py/requirements.txt
fastapi
uvicorn
playwright
GitPython
